requirements.txt
streamlit
scikit-learn
pandas
numpy
spacy
PyPDF2
python-docx

